+++ CHANGELOG.md
@@ -18,8 +18,7 @@ Features
 - Added informational message on startup if "surprise me on startup" is set. One if the color is changed. A different message if the color was not changed because one already existed.
 - Added setting for `showColorInStatusBar` setting (defaults to true).
 - Added three new elements that can be affected by Peacock's coloring. Each is colored with the same lighten/darken adjustments as activityBar
-  - `peacock.affectPanelBorder` Specifies whether Peacock should affect the panel's border (default false).
-  - `peacock.affectSideBarBorder` Specifies whether Peacock should affect the sideBar's border (default false).
+  - `peacock.affectAccentBorders` Specifies whether Peacock should affect the accent borders (panel, sidebar, editorGroup) Defaults to false.
   - `peacock.affectTabActiveBorder` Specifies whether Peacock should affect the active tab's border (default false)
 
 Security